Refactor search SQL to be contained in the search service objects.
This makes it easier to see how a search operates, and tweak
