+++ app/services/search_service.rb
@@ -1,8 +1,46 @@
 # frozen_string_literal: true
 
-require "./lib/sql/search"
-
 class SearchService
+  SEARCH_SQL = <<-SQL.squish
+    WITH sign_search(id, rank_precedence, rank_order, row_num) AS
+      (SELECT rs1.id,
+              rs1.rank_precedence,
+              rs1.rank_order,
+              ROW_NUMBER() OVER (PARTITION BY rs1.id
+                                ORDER BY rs1.rank_precedence ASC) AS row_num
+      FROM
+        (SELECT ranked_signs_by_exact_gloss.id,
+                1 AS rank_precedence,
+                RANK() OVER (ORDER BY ranked_signs_by_exact_gloss.word) AS rank_order
+          FROM signs ranked_signs_by_exact_gloss
+          WHERE UNACCENT(ranked_signs_by_exact_gloss.word) = :term
+          UNION SELECT ranked_signs_by_matched_gloss.id,
+                      2 AS rank_precedence,
+                      RANK() OVER (ORDER BY ranked_signs_by_matched_gloss.word) AS rank_order
+          FROM signs ranked_signs_by_matched_gloss
+          WHERE UNACCENT(ranked_signs_by_matched_gloss.word) ~* :term
+          UNION SELECT ranked_signs_by_exact_secondary.id,
+                      3 AS rank_precedence,
+                      RANK() OVER (ORDER BY ranked_signs_by_exact_secondary.secondary) AS rank_order
+          FROM signs ranked_signs_by_exact_secondary
+          WHERE UNACCENT(ranked_signs_by_exact_secondary.secondary) = :term
+          UNION SELECT ranked_signs_by_matched_secondary.id,
+                      4 AS rank_precedence,
+                      RANK() OVER (ORDER BY ranked_signs_by_matched_secondary.secondary) AS rank_order
+          FROM signs ranked_signs_by_matched_secondary
+          WHERE UNACCENT(ranked_signs_by_matched_secondary.secondary) ~* :term ) AS rs1)
+    SELECT *, activity.count AS activity_count
+    FROM sign_search
+    LEFT JOIN LATERAL (SELECT sign_activities.sign_id,
+                              COUNT(sign_activities.sign_id) AS count
+                      FROM sign_activities
+                      WHERE sign_search.row_num = 1 AND
+                            sign_activities.sign_id = sign_search.id AND
+                            sign_activities.key = 'agree'
+                      GROUP BY sign_activities.sign_id) AS activity ON TRUE
+    WHERE sign_search.row_num = 1
+  SQL
+
   attr_reader :search, :results
 
   def initialize(search:, relation:)
@@ -49,6 +87,6 @@ class SearchService
   end
 
   def prepare_search(term)
-    [SQL::Search.search(**search_args), { term: term }]
+    [self.class::SEARCH_SQL, { term: term }]
   end
 end
